requirements.txt
fitz==0.0.1.dev2
Flask==3.1.0
langchain==0.3.17
markdown2==2.5.2
openai==1.61.0
Pillow==11.1.0
pylatexenc==2.10
PyPDF2==3.0.1
transformers==4.48.0
frontend
tools
gunicorn
PyMuPDF
torch==2.0.1
